pyproject.toml
[tool.poetry]
name = "textgrab"
version = "0.1.0"
description = ""
authors = ["PowerSnail <hj@powersnail.com>"]
license = "GPLv2"

[tool.poetry.dependencies]
python = "~3.9"
pytesseract = "^0.3.7"
PySide6 = "^6.1.0"
docopt-ng = "^0.7.2"

[tool.poetry.dev-dependencies]
black = "^21.5b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
exclude = "ui_.*\\.py"

